services/sentiment_analysis_docker/python_requirements.txt
pandas
numpy
google-cloud-storage
pyarrow
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
transformers
emoji